@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file include/hrpc/RemoteKeeper.hpp
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,13 +27,13 @@
  *
  * @section DESCRIPTION
  *
- *  RemoteKeeper.hpp :  Help Handler
+ *  RemoteKeeper.hpp : Hrpc Remote Manager List Headers
  *
  */
 #ifndef _ZPDS_HRPC_REMOTE_KEEPER_HPP_
 #define _ZPDS_HRPC_REMOTE_KEEPER_HPP_
 
-#include <hrpc/HrpcClient.hpp>
+#include "hrpc/HrpcClient.hpp" // for headers
 
 namespace zpds {
 namespace hrpc {
@@ -123,7 +123,7 @@ public:
 	* MasterUpdate: elect the master based on info received after cutoff
 	*
 	* @param state
-	*   uint64_t cutoff 
+	*   uint64_t cutoff
 	*
 	* @return
 	*   none
