@@ -6,7 +6,7 @@
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  HttpServer.hpp :  HTTP Server  ( Modified from https://github.com/eidheim/Simple-Web-Server )
+ *  HttpServer.hpp : HTTP Server ( Modified from https://github.com/eidheim/Simple-Web-Server )
  *
  */
 #ifndef _ZPDS_HTTP_SERVER_HPP_
