@@ -6,7 +6,7 @@
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  HttpServerBase.hpp :  HTTP Server Base ( Modified from eidheim/Simple-Web-Server )
+ *  HttpServerBase.hpp : HTTP Server Base ( Modified from eidheim/Simple-Web-Server )
  *
  */
 #ifndef _ZPDS_HTTP_SERVER_BASE_HPP_
@@ -132,10 +132,8 @@ public:
 
 		void send_on_delete(const std::function<void(const error_code &)> &callback = nullptr) noexcept
 		{
-			session->connection->set_timeout(timeout_content);
 			auto self = this->shared_from_this(); // Keep Response instance alive through the following async_write
 			asio::async_write(*session->connection->socket, *streambuf, [self, callback](const error_code &ec, std::size_t /*bytes_transferred*/) {
-				self->session->connection->cancel_timeout();
 				auto lock = self->session->connection->handler_runner->continue_lock();
 				if(!lock)
 					return;
@@ -153,16 +151,14 @@ public:
 		/// Send the content of the response stream to client. The callback is called when the send has completed.
 		///
 		/// Use this function if you need to recursively send parts of a longer message, or when using server-sent events.
-		void send(const std::function<void(const error_code &)> &callback = nullptr) noexcept
+		void send(std::function<void(const error_code &)> callback = nullptr) noexcept
 		{
-			session->connection->set_timeout(timeout_content);
-
 			std::shared_ptr<asio::streambuf> streambuf = std::move(this->streambuf);
 			this->streambuf = std::unique_ptr<asio::streambuf>(new asio::streambuf());
 			rdbuf(this->streambuf.get());
 
 			LockGuard lock(send_queue_mutex);
-			send_queue.emplace_back(streambuf, callback);
+			send_queue.emplace_back(std::move(streambuf), std::move(callback));
 			if(send_queue.size() == 1)
 				send_from_queue();
 		}
@@ -234,19 +230,10 @@ public:
 		{
 			return streambuf.size();
 		}
-		/// Convenience function to return content as std::string. The stream buffer is consumed.
+		/// Convenience function to return content as std::string.
 		std::string string() noexcept
 		{
-			try {
-				std::string str;
-				auto size = streambuf.size();
-				str.resize(size);
-				read(&str[0], static_cast<std::streamsize>(size));
-				return str;
-			}
-			catch(...) {
-				return std::string();
-			}
+			return std::string(asio::buffers_begin(streambuf.data()), asio::buffers_end(streambuf.data()));
 		}
 
 	private:
@@ -390,13 +377,13 @@ public:
 		/// Port number to use. Defaults to 80 for HTTP and 443 for HTTPS. Set to 0 get an assigned port.
 		unsigned short port;
 		/// If io_service is not set, number of threads that the server will use when start() is called.
-		/// Timeout on request handling. Defaults to 5 seconds.
+		/// Timeout on request completion. Defaults to 5 seconds.
 		long timeout_request = 5;
-		/// Timeout on content handling. Defaults to 300 seconds.
+		/// Timeout on request/response content completion. Defaults to 300 seconds.
 		long timeout_content = 300;
 		/// Maximum size of request stream buffer. Defaults to architecture maximum.
 		/// Reaching this limit will result in a message_size error code.
-		std::size_t max_request_streambuf_size = std::numeric_limits<std::size_t>::max();
+		std::size_t max_request_streambuf_size = (std::numeric_limits<std::size_t>::max)();
 		/// IPv4 address in dotted decimal form or IPv6 address in hexadecimal notation.
 		/// If empty, the address will be any address.
 		std::string address;
@@ -438,12 +425,12 @@ public:
 	/// If you want to reuse an already created asio::io_service, store its pointer here before calling start().
 	std::shared_ptr<::zpds::http::io_whatever> io_whatever;
 
-	/// If you know the server port in advance, use start() instead.
-	/// Returns assigned port.
-	/// Call before accept_and_run().
-	unsigned short bind()
+	/// Start the server.
+	/// The callback argument is called after the server is accepting connections,
+	/// where its parameter contains the assigned port.
+	void start(const std::function<void(unsigned short /*port*/)> &callback = nullptr)
 	{
-		std::lock_guard<std::mutex> lock(start_stop_mutex);
+		std::unique_lock<std::mutex> lock(start_stop_mutex);
 
 		asio::ip::tcp::endpoint endpoint;
 		if(config.address.size() > 0)
@@ -463,25 +450,20 @@ public:
 #endif // End Linux
 		}
 		acceptor->bind(endpoint);
-
 		after_bind();
+		auto port = acceptor->local_endpoint().port();
 
-		return acceptor->local_endpoint().port();
-	}
-
-	/// If you know the server port in advance, use start() instead.
-	/// Call after bind().
-	void accept_and_run()
-	{
 		acceptor->listen();
 		accept();
-	}
 
-	/// Start the server by calling bind() and accept_and_run()
-	void start()
-	{
-		bind();
-		accept_and_run();
+		if(callback) {
+#if(BOOST_ASIO_VERSION >= 101300)
+			post(*io_whatever, [callback, port] { callback(port); });
+#else
+			io_whatever->post([callback, port] { callback(port); });
+#endif
+		}
+
 	}
 
 	/// Stop accepting new requests, and close current connections.
@@ -499,7 +481,6 @@ public:
 					connection->close();
 				connections->set.clear();
 			}
-
 		}
 	}
 
@@ -551,18 +532,11 @@ protected:
 	{
 		session->connection->set_timeout(config.timeout_request);
 		asio::async_read_until(*session->connection->socket, session->request->streambuf, "\r\n\r\n", [this, session](const error_code &ec, std::size_t bytes_transferred) {
-			session->connection->cancel_timeout();
+			session->connection->set_timeout(config.timeout_content);
 			auto lock = session->connection->handler_runner->continue_lock();
 			if(!lock)
 				return;
 			session->request->header_read_time = std::chrono::system_clock::now();
-			if(session->request->streambuf.size() == session->request->streambuf.max_size()) {
-				auto response = std::shared_ptr<Response>(new Response(session, this->config.timeout_content));
-				response->write(StatusCode::client_error_payload_too_large);
-				if(this->on_error)
-					this->on_error(session->request, make_error_code::make_error_code(errc::message_size));
-				return;
-			}
 
 			if(!ec) {
 				// request->streambuf.size() is not necessarily the same as bytes_transferred, from Boost-docs:
@@ -590,20 +564,18 @@ protected:
 							this->on_error(session->request, make_error_code::make_error_code(errc::protocol_error));
 						return;
 					}
+					if(content_length > session->request->streambuf.max_size()) {
+						auto response = std::shared_ptr<Response>(new Response(session, this->config.timeout_content));
+						response->write(StatusCode::client_error_payload_too_large);
+						if(this->on_error)
+							this->on_error(session->request, make_error_code::make_error_code(errc::message_size));
+						return;
+					}
 					if(content_length > num_additional_bytes) {
-						session->connection->set_timeout(config.timeout_content);
 						asio::async_read(*session->connection->socket, session->request->streambuf, asio::transfer_exactly(content_length - num_additional_bytes), [this, session](const error_code &ec, std::size_t /*bytes_transferred*/) {
-							session->connection->cancel_timeout();
 							auto lock = session->connection->handler_runner->continue_lock();
 							if(!lock)
 								return;
-							if(session->request->streambuf.size() == session->request->streambuf.max_size()) {
-								auto response = std::shared_ptr<Response>(new Response(session, this->config.timeout_content));
-								response->write(StatusCode::client_error_payload_too_large);
-								if(this->on_error)
-									this->on_error(session->request, make_error_code::make_error_code(errc::message_size));
-								return;
-							}
 
 							if(!ec)
 								this->find_resource(session);
@@ -615,16 +587,16 @@ protected:
 						this->find_resource(session);
 				}
 				else if((header_it = session->request->header.find("Transfer-Encoding")) != session->request->header.end() && header_it->second == "chunked") {
-					auto chunks_streambuf = std::make_shared<asio::streambuf>(this->config.max_request_streambuf_size);
+					// Expect hex number to not exceed 16 bytes (64-bit number), but take into account previous additional read bytes
+					auto chunk_size_streambuf = std::make_shared<asio::streambuf>(std::max<std::size_t>(16 + 2, session->request->streambuf.size()));
 
-					// Copy leftover bytes
-					std::ostream ostream(chunks_streambuf.get());
-					auto size = session->request->streambuf.size();
-					std::unique_ptr<char[]> buffer(new char[size]);
-					session->request->content.read(buffer.get(), static_cast<std::streamsize>(size));
-					ostream.write(buffer.get(), static_cast<std::streamsize>(size));
+					// Move leftover bytes
+					auto &source = session->request->streambuf;
+					auto &target = *chunk_size_streambuf;
+					target.commit(asio::buffer_copy(target.prepare(source.size()), source.data()));
+					source.consume(source.size());
 
-					this->read_chunked_transfer_encoded(session, chunks_streambuf);
+					this->read_chunked_transfer_encoded(session, chunk_size_streambuf);
 				}
 				else
 					this->find_resource(session);
@@ -634,31 +606,21 @@ protected:
 		});
 	}
 
-	void read_chunked_transfer_encoded(const std::shared_ptr<Session> &session, const std::shared_ptr<asio::streambuf> &chunks_streambuf)
+	void read_chunked_transfer_encoded(const std::shared_ptr<Session> &session, const std::shared_ptr<asio::streambuf> &chunk_size_streambuf)
 	{
-		session->connection->set_timeout(config.timeout_content);
-		asio::async_read_until(*session->connection->socket, *chunks_streambuf, "\r\n", [this, session, chunks_streambuf](const error_code &ec, size_t bytes_transferred) {
-			session->connection->cancel_timeout();
+		asio::async_read_until(*session->connection->socket, *chunk_size_streambuf, "\r\n", [this, session, chunk_size_streambuf](const error_code &ec, size_t bytes_transferred) {
 			auto lock = session->connection->handler_runner->continue_lock();
 			if(!lock)
 				return;
-			if(chunks_streambuf->size() == chunks_streambuf->max_size()) {
-				auto response = std::shared_ptr<Response>(new Response(session, this->config.timeout_content));
-				response->write(StatusCode::client_error_payload_too_large);
-				if(this->on_error)
-					this->on_error(session->request, make_error_code::make_error_code(errc::message_size));
-				return;
-			}
 
 			if(!ec) {
-				std::istream istream(chunks_streambuf.get());
+				std::istream istream(chunk_size_streambuf.get());
 				std::string line;
 				getline(istream, line);
 				bytes_transferred -= line.size() + 1;
-				line.pop_back();
-				unsigned long length = 0;
+				unsigned long chunk_size = 0;
 				try {
-					length = stoul(line, 0, 16);
+					chunk_size = stoul(line, 0, 16);
 				}
 				catch(...) {
 					if(this->on_error)
@@ -666,62 +628,61 @@ protected:
 					return;
 				}
 
-				auto num_additional_bytes = chunks_streambuf->size() - bytes_transferred;
+				if(2 + chunk_size + session->request->streambuf.size() > session->request->streambuf.max_size()) {
+					auto response = std::shared_ptr<Response>(new Response(session, this->config.timeout_content));
+					response->write(StatusCode::client_error_payload_too_large);
+					if(this->on_error)
+						this->on_error(session->request, make_error_code::make_error_code(errc::message_size));
+					return;
+				}
 
-				if((2 + length) > num_additional_bytes) {
-					session->connection->set_timeout(config.timeout_content);
-					asio::async_read(*session->connection->socket, *chunks_streambuf, asio::transfer_exactly(2 + length - num_additional_bytes), [this, session, chunks_streambuf, length](const error_code &ec, size_t /*bytes_transferred*/) {
-						session->connection->cancel_timeout();
+				auto num_additional_bytes = chunk_size_streambuf->size() - bytes_transferred;
+
+				auto bytes_to_move = std::min<std::size_t>(chunk_size, num_additional_bytes);
+				if(bytes_to_move > 0) {
+					// Move leftover bytes
+					auto &source = *chunk_size_streambuf;
+					auto &target = session->request->streambuf;
+					target.commit(asio::buffer_copy(target.prepare(bytes_to_move), source.data(), bytes_to_move));
+					source.consume(bytes_to_move);
+				}
+
+				if((2 + chunk_size) > num_additional_bytes) {
+					asio::async_read(*session->connection->socket, session->request->streambuf, asio::transfer_exactly(2 + chunk_size - num_additional_bytes), [this, session, chunk_size_streambuf, chunk_size](const error_code &ec, size_t /*bytes_transferred*/) {
 						auto lock = session->connection->handler_runner->continue_lock();
 						if(!lock)
 							return;
-						if(chunks_streambuf->size() == chunks_streambuf->max_size()) {
-							auto response = std::shared_ptr<Response>(new Response(session, this->config.timeout_content));
-							response->write(StatusCode::client_error_payload_too_large);
-							if(this->on_error)
-								this->on_error(session->request, make_error_code::make_error_code(errc::message_size));
-							return;
-						}
 
-						if(!ec)
-							this->read_chunked_transfer_encoded_chunk(session, chunks_streambuf, length);
+						if(!ec) {
+							std::istream istream(&session->request->streambuf);
+
+							// Remove "\r\n"
+							istream.get();
+							istream.get();
+
+							if(chunk_size > 0)
+								read_chunked_transfer_encoded(session, chunk_size_streambuf);
+							else
+								this->find_resource(session);
+						}
 						else if(this->on_error)
 							this->on_error(session->request, ec);
 					});
 				}
-				else
-					this->read_chunked_transfer_encoded_chunk(session, chunks_streambuf, length);
+				else {
+					// Remove "\r\n"
+					istream.get();
+					istream.get();
+
+					if(chunk_size > 0)
+						read_chunked_transfer_encoded(session, chunk_size_streambuf);
+					else
+						this->find_resource(session);
+				}
 			}
 			else if(this->on_error)
 				this->on_error(session->request, ec);
 		});
-	}
-
-	void read_chunked_transfer_encoded_chunk(const std::shared_ptr<Session> &session, const std::shared_ptr<asio::streambuf> &chunks_streambuf, unsigned long length)
-	{
-		std::istream istream(chunks_streambuf.get());
-		if(length > 0) {
-			std::ostream ostream(&session->request->streambuf);
-			std::unique_ptr<char[]> buffer(new char[length]);
-			istream.read(buffer.get(), static_cast<std::streamsize>(length));
-			ostream.write(buffer.get(), static_cast<std::streamsize>(length));
-			if(session->request->streambuf.size() == session->request->streambuf.max_size()) {
-				auto response = std::shared_ptr<Response>(new Response(session, this->config.timeout_content));
-				response->write(StatusCode::client_error_payload_too_large);
-				if(this->on_error)
-					this->on_error(session->request, make_error_code::make_error_code(errc::message_size));
-				return;
-			}
-		}
-
-		// Remove "\r\n"
-		istream.get();
-		istream.get();
-
-		if(length > 0)
-			read_chunked_transfer_encoded(session, chunks_streambuf);
-		else
-			this->find_resource(session);
 	}
 
 	void find_resource(const std::shared_ptr<Session> &session)
@@ -762,10 +723,10 @@ protected:
 	void write(const std::shared_ptr<Session> &session,
 	           std::function<void(std::shared_ptr<typename HttpServerBase<socket_type>::Response>, std::shared_ptr<typename HttpServerBase<socket_type>::Request>)> &resource_function)
 	{
-		session->connection->set_timeout(config.timeout_content);
 		auto response = std::shared_ptr<Response>(new Response(session, config.timeout_content), [this](Response *response_ptr) {
 			auto response = std::shared_ptr<Response>(response_ptr);
 			response->send_on_delete([this, response](const error_code &ec) {
+				response->session->connection->cancel_timeout();
 				if(!ec) {
 					if(response->close_connection_after_response)
 						return;
