@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file include/jamspell/BloomFilter.hpp
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  BloomFilter.hpp : Bllom filter headers for jamspell
+ *  BloomFilter.hpp : Jamspell Bloom filter Headers for jamspell
  *
  */
 #ifndef _ZPDS_JAMSPELL_BLOOM_FILTER_HPP_
