@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file include/query/CategoryService.hpp
- * @author  S Roychowdhury <sroycode AT gmail DOT com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury.
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  CategoryService.hpp : Data input service header
+ *  CategoryService.hpp : Category Manipulation Endpoint
  *
  */
 #ifndef _ZPDS_QUERY_CATEGORY_SERVICE_HPP_
@@ -35,10 +35,11 @@
 
 #include "query/QueryBase.hpp"
 #include "hrpc/HrpcClient.hpp"
-#include "store/ProfileService.hpp"
+
 #include "store/CategoryService.hpp"
-#include <boost/lexical_cast.hpp>
-#include <boost/algorithm/string.hpp>
+
+#include <google/protobuf/descriptor.h>
+#include <google/protobuf/message.h>
 
 namespace zpds {
 namespace query {
@@ -75,21 +76,55 @@ public:
 	{
 		if (!(this->myscope & scope)) return; // scope mismatch
 
-		// Endpoint : POST _data/api/v1/category/upsert
-		helpquery->add({scope, "POST _data/api/v1/category/upsert", {
-				"Update and overwrite",
-				"Parameters: see CatDataT"
+		// Endpoint : POST _admin/api/v1/category/create
+
+		helpquery->add({scope, "POST _admin/api/v1/category/create", {
+				"create hosts and related data",
+				"Parameters:",
+				"username : updater name",
+				"sessionkey: updater session key",
+				"payload: contains payload for category, name(*), lang(def: EN), description ",
+				"payloads: array of above"
 			}
 		});
 
-		// Endpoint : POST _data/api/v1/category/merge
-		helpquery->add({scope, "POST _data/api/v1/category/merge", {
-				"Update and merge",
-				"Parameters: see CatDataT"
+		// Endpoint : POST _admin/api/v1/category/upsert
+
+		helpquery->add({scope, "POST _admin/api/v1/category/upsert", {
+				"upsert hosts and related data",
+				"Parameters:",
+				"username : updater name",
+				"sessionkey: updater session key",
+				"payload: contains payload for category, name(*), lang(def: EN), description",
+				"payloads: array of above"
 			}
 		});
 
-		server->resource["/_data/api/v1/category/(upsert|merge)$"]["POST"]
+		// Endpoint : POST _admin/api/v1/category/update
+
+		helpquery->add({scope, "POST _admin/api/v1/category/update", {
+				"update hosts and related data",
+				"Parameters:",
+				"username : updater name",
+				"sessionkey: updater session key",
+				"payload: contains payload for category, name(*), lang(def: EN), description",
+				"payloads: array of above"
+			}
+		});
+
+		// Endpoint : POST _admin/api/v1/category/delete
+
+		helpquery->add({scope, "POST _admin/api/v1/category/delete", {
+				"delete hosts and related data",
+				"Parameters:",
+				"username : updater name",
+				"sessionkey: updater session key",
+				"payload: contains payload for category, name(*)",
+				"payloads: array of above"
+			}
+		});
+
+		server->resource["/_admin/api/v1/category/(create|upsert|update|delete)$"]["POST"]
 		=[this,stptr](typename HttpServerT::RespPtr response, typename HttpServerT::ReqPtr request) {
 			ZPDS_PARALLEL_ONE([this,stptr,response,request] {
 				uint64_t currtime = ZPDS_CURRTIME_MS;
@@ -98,61 +133,57 @@ public:
 				try
 				{
 					bool use_json = this->JsonRequest(request);
-
-					zpds::query::CatParamsT lparams;
-					auto ldata = lparams.mutable_ldata();
-
+					zpds::query::CategoryRespT getp;
 					if (use_json) {
-						bool stat = json2pb( request->content.string(), ldata );
-						if (!stat) throw InitialException("Bad Json Format",M_BAD_JSON);
+						int stat = json2pb( request->content.string(), &getp);
+						if (stat<0) throw InitialException("Bad Json Format",M_BAD_JSON);
 					}
 					else {
 						throw InitialException("Only Json Format",M_BAD_JSON);
 					}
 
-					ldata->set_merge(  request->path_match[1]=="merge" );
+					if (getp.username().empty())
+						throw InitialException("Bad Json Data",M_BAD_JSON);
 
-					if (ldata->records_size() > ZPDS_UPLOAD_MAX_RECORDS) 
-							throw zpds::BadDataException("Cannot handle that many records, max exceeded",M_INVALID_PARAM);
+					if (getp.payloads_size() > ZPDS_NUPLOAD_MAX_RECORDS)
+						throw zpds::BadDataException("Cannot handle that many records, max exceeded",M_INVALID_PARAM);
+
+					// set action
+					if (request->path_match[1]=="create" ) getp.set_action(::zpds::query::W_CREATE);
+					else if (request->path_match[1]=="upsert" ) getp.set_action(::zpds::query::W_UPSERT);
+					else if (request->path_match[1]=="update" ) getp.set_action(::zpds::query::W_UPDATE);
+					else if (request->path_match[1]=="delete" ) getp.set_action(::zpds::query::W_DELETE);
+					else throw zpds::BadCodeException("Unknown API used");
 
 					if (!stptr->is_ready.Get()) throw zpds::BadDataException("System Not Ready");
 
 					// action
-					if (!stptr->allow_free_update.Get()) {
-						::zpds::store::ProfileService pf;
-						::zpds::store::ProfileT prof;
-						prof.set_name( ldata->name() );
-						prof.set_passkey( ldata->passkey() );
-						if (!pf.VerifyProfile(stptr,&prof,true) )
-							throw zpds::BadDataException("No such user or Invalid passkey or user cannot update",M_INVALID_PARAM);
-					}
-					ldata->clear_passkey();
-
 					if (stptr->is_master.Get()) {
 						::zpds::store::CategoryService rs;
-						rs.AddDataAction(stptr,&lparams);
+						rs.ManageDataAction(stptr,&getp);
 					}
 					else {
 						zpds::hrpc::HrpcClient hclient;
-						hclient.SendToMaster(stptr,::zpds::hrpc::M_ADDDATA_CAT,&lparams);
+						hclient.SendToMaster(stptr,::zpds::hrpc::M_MANAGE_CATEGORY,&getp);
 					}
+
 					// aftermath
 
 					std::string output;
-					pb2json(lparams.mutable_lresp(),output);
+					pb2json(getp.mutable_status(),output);
 
 					ok=true;
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (zpds::BaseException& e)
 				{
 					std::string output = err2json(e.ecode(),e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (std::exception& e)
 				{
 					std::string output = err2json(M_UNKNOWN,e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (...)
 				{
@@ -162,23 +193,31 @@ public:
 			});
 		};
 
-		// Endpoint : POST _data/api/v1/category/get
+		// Endpoint : POST _admin/api/v1/category/getone
 
-		helpquery->add({scope, "POST _data/api/v1/category/get", {
-				"gets the data as json",
-				"Parameters: see CatDataT, just need stype lang uniqueid"
+		helpquery->add({scope, "POST _admin/api/v1/category/getone", {
+				"for exter users to read category and related data",
+				"using name(*)",
+				"Parameters:",
+				"username : reader name",
+				"sessionkey: reader session key",
+				"payload: contains payload for name"
 			}
 		});
 
-		// Endpoint : POST _data/api/v1/category/getnd
+		// Endpoint : POST _admin/api/v1/category/getmany
 
-		helpquery->add({scope, "POST _data/api/v1/category/getnd", {
-				"gets the data as ndjson",
-				"Parameters: see CatDataT, just need stype lang uniqueid"
+		helpquery->add({scope, "POST _admin/api/v1/category/getmany", {
+				"for exter users to read category and related data",
+				"using name(*)",
+				"Parameters:",
+				"username : reader name",
+				"sessionkey: reader session key",
+				"payloads: contains array of payload for name"
 			}
 		});
 
-		server->resource["/_data/api/v1/category/get(nd|)$"]["POST"]
+		server->resource["/_admin/api/v1/category/(getone|getmany)$"]["POST"]
 		=[this,stptr](typename HttpServerT::RespPtr response, typename HttpServerT::ReqPtr request) {
 			ZPDS_PARALLEL_ONE([this,stptr,response,request] {
 				uint64_t currtime = ZPDS_CURRTIME_MS;
@@ -187,54 +226,60 @@ public:
 				try
 				{
 					bool use_json = this->JsonRequest(request);
-					bool use_nd = (request->path_match[1]=="nd");
-
-					zpds::query::CatParamsT lparams;
-					auto ldata = lparams.mutable_ldata();
-
+					zpds::query::CategoryRespT getp;
 					if (use_json) {
-						bool stat = json2pb( request->content.string(), ldata );
-						if (!stat) throw InitialException("Bad Json Format",M_BAD_JSON);
+						int stat = json2pb( request->content.string(), &getp);
+						if (stat<0) throw InitialException("Bad Json Format",M_BAD_JSON);
 					}
 					else {
 						throw InitialException("Only Json Format",M_BAD_JSON);
 					}
 
-					// extra
-					lparams.set_extra( ldata->extra() > 0 );
+					if (getp.username().empty())
+						throw InitialException("Bad Json Data",M_BAD_JSON);
+
+					if (getp.payloads_size() > ZPDS_NUPLOAD_MAX_RECORDS)
+						throw zpds::BadDataException("Cannot handle that many records, max exceeded",M_INVALID_PARAM);
+
+					DLOG(INFO) << getp.DebugString() << std::endl;
+
+					// set action
+					if (request->path_match[1]=="getone" ) {
+						getp.clear_payloads();
+					}
+					else if (request->path_match[1]=="getmany" ) {
+						getp.clear_payload();
+					}
+					else throw zpds::BadCodeException("Unknown API used");
 
 					if (!stptr->is_ready.Get()) throw zpds::BadDataException("System Not Ready");
 
 					// action
-
 					::zpds::store::CategoryService rs;
-					rs.GetDataAction(stptr,&lparams);
+					rs.ReadDataAction(stptr,&getp);
+
 					// aftermath
+
 					std::string output;
-					if (use_nd) {
-						std::stringstream ss;
-						for (auto i=0; i<ldata->records_size(); ++i) {
-							if (!ldata->records(i).notfound()) // remove notfound
-								ss << pb2json(ldata->mutable_records(i)) << "\n";
-						}
-						output = ss.str();
-					}
-					else {
-						pb2json(ldata,output);
-					}
+					// clear getp fields
+					getp.clear_username();
+					getp.clear_sessionkey();
+					getp.clear_status();
+					getp.clear_action();
+					pb2json(&getp,output);
 
 					ok=true;
-					this->HttpOKAction(response,request,200,"OK", (use_nd) ? "application/x-ndjson" : "application/json", output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (zpds::BaseException& e)
 				{
 					std::string output = err2json(e.ecode(),e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (std::exception& e)
 				{
 					std::string output = err2json(M_UNKNOWN,e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (...)
 				{
@@ -244,128 +289,9 @@ public:
 			});
 		};
 
-		// Endpoint : GET _data/api/v1/category/esbulk
-
-		helpquery->add({scope, "GET _data/api/v1/category/esbulk?{params}", {
-				"this gets the elastic data",
-				"Parameters:",
-				"index: _index to populate",
-				"type: _type to populate",
-				"styp: record type to get",
-				"lang: language type to get",
-				"last_up: updated after",
-				"last_id: id of last updated optional",
-				"limit: limit optional default 1000 ,max 10k"
-			}
-		});
-
-		server->resource["/_data/api/v1/category/esbulk$"]["GET"]
-		=[this,stptr](typename HttpServerT::RespPtr response, typename HttpServerT::ReqPtr request) {
-			ZPDS_PARALLEL_ONE([this,stptr,response,request] {
-				uint64_t currtime = ZPDS_CURRTIME_MS;
-				bool ok=false;
-
-				try
-				{
-					auto params = urldecode( request->query_string );
-					zpds::query::CatParamsT lparams;
-
-					// index for es
-					if ( params.find("index") == params.end() )
-						throw zpds::InitialException("index variable needed",M_INVALID_PARAM);
-					std::string e_index{params["index"]};
-
-					// type for es
-					if ( params.find("type") == params.end() )
-						throw zpds::InitialException("type variable needed",M_INVALID_PARAM);
-					std::string e_type{params["type"]};
-
-					// styp default S_NONE
-					if ( params.find("styp") == params.end() ) {
-						const google::protobuf::EnumDescriptor *descriptor = zpds::search::SourceTypeE_descriptor();
-						if ( descriptor->FindValueByName( params["styp"] ) )
-							lparams.set_styp( zpds::search::SourceTypeE ( descriptor->FindValueByName( params["styp"])->number() ) );
-					}
-
-					// lang default EN
-					if ( params.find("lang") == params.end() ) {
-
-						std::string lang = params["lang"];
-						boost::algorithm::to_upper(lang);
-						const google::protobuf::EnumDescriptor *descriptor = zpds::search::LangTypeE_descriptor();
-						if ( descriptor->FindValueByName( lang ) )
-							lparams.set_lang( zpds::search::LangTypeE ( descriptor->FindValueByName(lang)->number() ) );
-					}
-
-					// updated_at
-					if ( params.find("last_up") != params.end() ) {
-						if (params["last_up"].length()>0)
-							lparams.set_last_up( boost::lexical_cast<uint64_t>(params["last_up"]) );
-					}
-
-					// last_id
-					if ( params.find("last_id") != params.end() ) {
-						if (params["last_id"].length()>0)
-							lparams.set_last_id( boost::lexical_cast<uint64_t>(params["last_id"]) );
-					}
-
-					// limit
-					lparams.set_limit( 1000 ); // default limit
-					if ( params.find("limit") != params.end() ) {
-						if (params["limit"].length()>0)
-							lparams.set_limit( boost::lexical_cast<uint64_t>(params["limit"]) );
-					}
-
-					// extra
-					if ( params.find("extra") != params.end() ) {
-						if (params["extra"].length()>0)
-							lparams.set_extra( boost::lexical_cast<uint64_t>(params["last_id"]) > 0 );
-					}
-
-					if (!stptr->is_ready.Get()) throw zpds::BadDataException("System Not Ready");
-
-					::zpds::store::CategoryService rs;
-					rs.GetIndexDataAction(stptr,&lparams);
-					// aftermath
-					auto ldata = lparams.mutable_ldata();
-					std::stringstream ss;
-					for (auto i=0; i<ldata->records_size(); ++i) {
-						if (ldata->records(i).notfound()) continue;
-						if (ldata->records(i).is_deleted()) {
-							ss << "{ \"delete\" : { \"_index\" : \"" << e_index
-							   << "\", \"_type\" : \"" << e_type
-							   << "\", \"_id\" : \"" << ldata->records(i).id() << "\" } }\n";
-						}
-						else {
-							ss << "{ \"index\" : { \"_index\" : \"" << e_index
-							   << "\", \"_type\" : \"" << e_type
-							   << "\", \"_id\" : \"" << ldata->records(i).id() << "\" } }\n"
-							   << pb2json(ldata->mutable_records(i)) << "\n";
-						}
-					}
-					std::string output = ss.str();
-					ok=true;
-					this->HttpOKAction(response,request,200,"OK","application/x-ndjson",output,true);
-				}
-				catch (zpds::BaseException& e)
-				{
-					std::string output = err2json(e.ecode(),e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
-				}
-				catch (std::exception& e)
-				{
-					std::string output = err2json(M_UNKNOWN,e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
-				}
-				catch (...)
-				{
-					this->HttpErrorAction(response,request,500,"INTERNAL SERVER ERROR");
-				}
-				LOG(INFO) << request->path << " ms: " << ZPDS_CURRTIME_MS - currtime;
-			});
-		};
-
+		// end
 	}
+
 
 private:
 
@@ -374,4 +300,5 @@ private:
 } // namespace zpds
 
 #endif // _ZPDS_QUERY_CATEGORY_SERVICE_HPP_
+
 
