@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file include/query/CommandService.hpp
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,16 +27,14 @@
  *
  * @section DESCRIPTION
  *
- *  CommandService.hpp : Command related services
+ *  CommandService.hpp : Command related services Endpoint
  *
  */
 #ifndef _ZPDS_QUERY_COMMAND_SERVICE_HPP_
 #define _ZPDS_QUERY_COMMAND_SERVICE_HPP_
 
 #include "query/QueryBase.hpp"
-
-#include <google/protobuf/descriptor.h>
-#include <google/protobuf/message.h>
+#include "store/ExterCredService.hpp"
 
 namespace zpds {
 namespace query {
@@ -76,9 +74,20 @@ public:
 		// Endpoint : GET _admin/api/v1/local_command/{command}
 
 		helpquery->add({scope, "GET _admin/api/v1/local_command/commit", {
-				"this forces indexing "
+				"this forces indexing at next update"
 			}
 		});
+
+		helpquery->add({scope, "GET _admin/api/v1/local_command/commitnow", {
+				"this forces indexing imediately"
+			}
+		});
+
+		helpquery->add({scope, "GET _admin/api/v1/local_command/noxapian", {
+				"this starts ignoring search data in update"
+			}
+		});
+
 
 		server->resource["/_admin/api/v1/local_command/(.*)$"]["GET"]
 		=[this,stptr](typename HttpServerT::RespPtr response, typename HttpServerT::ReqPtr request) {
@@ -96,14 +105,35 @@ public:
 					if (it->second != stptr->shared_secret.Get())
 						throw zpds::BadDataException("Bad Shared Secret");
 
-
 					if (!stptr->is_ready.Get()) throw zpds::BadDataException("System Not Ready");
 
 					::zpds::store::ErrorT errt;
 					// action
 					if ( request->path_match[1]=="commit" ) {
+#ifdef ZPDS_BUILD_WITH_XAPIAN
 						stptr->force_commit.Set(true);
 						errt.set_status("OK");
+#else
+						errt.set_status("NA");
+#endif
+					}
+
+					else if ( request->path_match[1]=="commitnow" ) {
+#ifdef ZPDS_BUILD_WITH_XAPIAN
+						stptr->xapdb->CommitData();
+						errt.set_status("OK");
+#else
+						errt.set_status("NA");
+#endif
+					}
+
+					else if ( request->path_match[1]=="noxapian" ) {
+#ifdef ZPDS_BUILD_WITH_XAPIAN
+						stptr->no_xapian.Set(true);
+						errt.set_status("OK");
+#else
+						errt.set_status("NA");
+#endif
 					}
 
 					// aftermath
@@ -111,17 +141,17 @@ public:
 					pb2json(&errt,output);
 
 					ok=true;
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (zpds::BaseException& e)
 				{
 					std::string output = err2json(e.ecode(),e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (std::exception& e)
 				{
 					std::string output = err2json(M_UNKNOWN,e.what());
-					this->HttpOKAction(response,request,200,"OK","application/json",output,true);
+					this->HttpOKAction(response,request,200,"OK","application/json",output);
 				}
 				catch (...)
 				{
