@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file include/search/DistanceSlabKeyMaker.hpp
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,20 +27,13 @@
  *
  * @section DESCRIPTION
  *
- *  DistanceSlabKeyMaker.hpp :   For Xapian Keys distance and values of 32-bit integer
+ *  DistanceSlabKeyMaker.hpp : For Xapian Keys distance and values of 32-bit integer
  *
  */
 #ifndef _ZPDS_SEARCH_DISTANCE_SLAB_KEYMAKER_HPP_
 #define _ZPDS_SEARCH_DISTANCE_SLAB_KEYMAKER_HPP_
 
-#include <cmath>
-#include <boost/format.hpp>
 #include <xapian.h>
-
-#include "utils/BaseUtils.hpp"
-#include "utils/StringHelpers.hpp"
-#include <vector>
-#include <limits>
 
 #define XAP_DSLAB_MAX_IMPORTANCE 100
 #define XAP_DSLAB_MAX_DISTANCE  400000
@@ -51,6 +44,7 @@ namespace search {
 class DistanceSlabKeyMaker : public Xapian::KeyMaker {
 
 public:
+
 	/**
 	* Constructor: default
 	*
@@ -79,24 +73,13 @@ public:
 	    const Xapian::LatLongCoords& centre_,
 	    int slab_,
 	    int dd= XAP_DSLAB_MAX_DISTANCE,
-	    int md= XAP_DSLAB_MAX_DISTANCE)
-		: slot(slot_),
-		  slot_o(slot_o_),
-		  centre(centre_),
-		  metric(new Xapian::GreatCircleMetric()),
-		  slab( KeepInBound<int>(slab_, 1, XAP_DSLAB_MAX_DISTANCE ) ),
-		  defdist( KeepInBound<int>(dd, 0, XAP_DSLAB_MAX_DISTANCE ) ),
-		  maxdist( KeepInBound<int>(md, 0, XAP_DSLAB_MAX_DISTANCE ) )
-	{}
+	    int md= XAP_DSLAB_MAX_DISTANCE);
 
 	/**
 	* Destructor: default
 	*
 	*/
-	~DistanceSlabKeyMaker()
-	{
-		delete metric;
-	}
+	~DistanceSlabKeyMaker();
 
 	/**
 	* operator(): default
@@ -107,27 +90,7 @@ public:
 	*  @return
 	*   std::string
 	*/
-	std::string operator()(const Xapian::Document& doc) const
-	{
-
-		// handle dist as int32 faster
-		int dist = defdist;
-		std::string val_l(doc.get_value(slot));
-		if (!val_l.empty()) {
-			Xapian::LatLongCoords doccoords;
-			doccoords.unserialise(val_l);
-			dist = KeepInBound<int>( std::lround((*metric)(centre, doccoords)) , 0, maxdist);
-		}
-		int dslab = 1+ (dist/slab);
-
-		// handle cc
-		double val_o = KeepInBound<double>( Xapian::sortable_unserialise(doc.get_value(slot_o)) , 0, XAP_DSLAB_MAX_IMPORTANCE );
-
-		// if (dist!=defdist)
-		// LOG(INFO) << "slab=" << slab << ", dist=" << dist << " , dslab=" << dslab
-		// << ", val_o=" << val_o << " , final=" << (double)(dslab * XAP_DSLAB_MAX_IMPORTANCE) + (XAP_DSLAB_MAX_IMPORTANCE-val_o);
-		return Xapian::sortable_serialise( (double)(dslab * XAP_DSLAB_MAX_IMPORTANCE) + (XAP_DSLAB_MAX_IMPORTANCE-val_o) );
-	}
+	std::string operator()(const Xapian::Document& doc) const;
 
 protected:
 	Xapian::valueno slot; // distance
