@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file include/utils/ServerBase.hpp
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  ServerBase.hpp :   Server ABC
+ *  ServerBase.hpp : Generic Abstract Base Class for all Servers Headers
  *
  */
 #ifndef _ZPDS_UTILS_SERVERBASE_HPP_
