@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file include/utils/SharedTable.hpp
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  SharedTable.hpp :   Shared pointer to sharedtable object
+ *  SharedTable.hpp : Shared Table of system data Headers
  *
  */
 #ifndef _ZPDS_UTILS_SHAREDTABLE_HPP_
@@ -37,15 +37,24 @@
 #include <vector>
 #include <random>
 
+#include <boost/asio.hpp>
+#include "http/AsioCompat.hpp"
+
 #include "utils/SharedCounter.hpp"
 #include "utils/SharedPairMap.hpp"
 #include "utils/SharedMap.hpp"
+// #include "utils/SharedQueue.hpp"
 
 #include "store/StoreLevel.hpp"
 
 #include "store/CacheContainer.hpp"
-#include "search/StoreTrie.hpp"
+
+#ifdef ZPDS_BUILD_WITH_XAPIAN
+#include "search/WriteIndex.hpp"
 #include "jamspell/StoreJam.hpp"
+#endif
+
+#include "crypto/CryptoBase.hpp"
 
 namespace zpds {
 namespace utils {
@@ -58,6 +67,7 @@ public:
 	using SharedUnsigned = SharedObject<uint64_t>;
 	using SharedDBPointer = SharedObject<dbpointer>;
 	using SharedBool = SharedObject<bool>;
+	// using JobQueue = SharedQueue<std::string>;
 
 	using SharedRemote = SharedPairMap<std::string,uint64_t,uint64_t>;
 	using RemoteMapT = SharedPairMap<std::string,uint64_t,uint64_t>::PairMapT;
@@ -65,10 +75,18 @@ public:
 	using SharedTrans = SharedMap<uint64_t,std::string>;
 
 	using SharedCache = zpds::store::CacheContainer::pointer;
-	using SharedTrie = zpds::search::StoreTrie::pointer;
-	using SharedJam = zpds::jamspell::StoreJam::pointer;
 
-	using LockT = boost::shared_mutex;
+#ifdef ZPDS_BUILD_WITH_XAPIAN
+	using SharedXap = zpds::search::WriteIndex::pointer;
+	using SharedJam = zpds::jamspell::StoreJam::pointer;
+#endif
+
+	using KeyRingT = std::unordered_map<std::string,std::shared_ptr<zpds::crypto::CryptoBase> >;
+
+	// using LockT = boost::shared_mutex;
+
+	// io_whatever
+	std::shared_ptr<::zpds::http::io_whatever> io_whatever;
 
 	// map of shared followers
 	SharedRemote remotes;
@@ -80,16 +98,24 @@ public:
 	SharedString master;
 	SharedString shared_secret;
 	SharedString hostname;
-	SharedString baseurl;
 	SharedString thisurl;
 	SharedString lastslave;
 
-	// whatwhere
+	// keyring
+	KeyRingT keyring;
+
+#ifdef ZPDS_BUILD_WITH_XAPIAN
+	// xapian
 	SharedString xapath;
-	SharedTrie xapdb;
+	SharedXap xapdb;
 
 	// spellcheck
 	SharedJam jamdb;
+
+	// xapian dont use
+	SharedBool no_xapian;
+
+#endif
 
 	// counter shared
 	SharedCounter maincounter;
@@ -101,20 +127,21 @@ public:
 
 	// cache pointers
 	SharedCache dbcache;
+	SharedCache tmpcache;
 
 	// booleans
 	SharedBool is_master;
 	SharedBool is_ready;
 
-	SharedBool dont_use_cache;
-	SharedBool dont_create_profile;
-	SharedBool allow_free_update;
-	SharedBool use_local_whatwhere;
-
 	SharedBool force_commit;
+	SharedBool lock_updates;
 
-	// lockers
-	LockT update_lock;
+	// shared
+	SharedUnsigned max_fetch_records;
+	SharedUnsigned max_user_sessions;
+
+	// queue
+	// JobQueue jobqueue;
 
 	/**
 	* make noncopyable
@@ -131,8 +158,7 @@ public:
 	*/
 	static pointer create()
 	{
-		pointer p(new SharedTable());
-		return p;
+		return pointer{new SharedTable};
 	}
 
 	/**
@@ -161,7 +187,8 @@ private:
 	*   none
 	*/
 	SharedTable() :
-		dbcache(zpds::store::CacheContainer::create())
+		dbcache(zpds::store::CacheContainer::create()),
+		tmpcache(zpds::store::CacheContainer::create())
 	{}
 
 };
