@@ -6,7 +6,7 @@
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  WebServer.hpp :   Web Server Headers
+ *  WebServer.hpp : Web Server Headers
  *
  */
 #ifndef _ZPDS_HTTP_WEBSERVER_HPP_
