@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file src/main/DefaultServer.hh
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,7 +27,7 @@
  *
  * @section DESCRIPTION
  *
- *  DefaultServer.hh :   DefaultServer Headers File
+ *  DefaultServer.hh : Default Server Headers File
  *
  */
 #ifndef _ZPDS_DEFAULTSERVER_HH_
@@ -35,7 +35,11 @@
 
 #define ZPDS_DEFAULT_EXE_NAME "zpds_server"
 #define ZPDS_DEFAULT_EXE_VERSION "1.0.0"
-#define ZPDS_DEFAULT_EXE_COPYRIGHT "Copyright (c) 2018-2019"
+#define ZPDS_DEFAULT_EXE_COPYRIGHT "Copyright (c) 2020 S Roychowdhury"
+
+#define ZPDS_DEFAULT_STRN_SYSTEM "system"
+#define ZPDS_DEFAULT_STRN_WORK   "work"
+#define ZPDS_DEFAULT_STRN_XAPIAN "xapian"
 
 #endif
 
