@@ -1,12 +1,12 @@
 /**
  * @project zapdos
  * @file src/tools/DumpDb.cc
- * @author  S Roychowdhury < sroycode at gmail dot com>
+ * @author  S Roychowdhury < sroycode at gmail dot com >
  * @version 1.0.0
  *
  * @section LICENSE
  *
- * Copyright (c) 2018-2019 S Roychowdhury
+ * Copyright (c) 2018-2020 S Roychowdhury
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -27,26 +27,46 @@
  *
  * @section DESCRIPTION
  *
- *  DumpDb.cc : DB Dumper
+ *  DumpDb.cc : zpds_dumpdb : Debugging DB Dumper
  *
  */
-
 #define ZPDS_DEFAULT_EXE_NAME "zpds_dumpdb"
 #define ZPDS_DEFAULT_EXE_VERSION "1.0.0"
-#define ZPDS_DEFAULT_EXE_COPYRIGHT "Copyright (c) 2018-2019 S Roychowdhury"
+#define ZPDS_DEFAULT_EXE_COPYRIGHT "Copyright (c) 2019 S Roychowdhury"
+
+#define ZPDS_CASE_MACRO(XXXSC,XXXUC) \
+case XXXUC:	{ XXXSC x; bool status = x.ParseFromString(it->value().ToString()); \
+if (status && x.id()>0) std::cout << #XXXUC << " : " << x.id() << " : " << pb2json(&x) << std::endl; break; }
 
 #include <iostream>
+#include <google/protobuf/descriptor.h>
+#include <google/protobuf/message.h>
+#include <google/protobuf/util/json_util.h>
 #include "store/StoreLevel.hpp"
 #include "utils/BaseUtils.hpp"
 
+
 namespace zpds {
 namespace store {
+
 class TablePrint : public StoreLevel {
 public:
 	using StoreLevel::dbpointer;
 	using StoreLevel::StoreLevel;
 
 	TablePrint() = delete;
+
+	std::string pb2json(const google::protobuf::Message* msg)
+	{
+		std::string str;
+		google::protobuf::util::JsonPrintOptions options;
+		options.add_whitespace = false;
+		options.preserve_proto_field_names = true;
+		// options.always_print_primitive_fields = true;
+		if (google::protobuf::util::MessageToJsonString(*msg, &str, options) != google::protobuf::util::Status::OK)
+			str="Broken Data";
+		return str;
+	}
 
 	void PrintAll()
 	{
@@ -56,37 +76,20 @@ public:
 			if (key.length() < ( ZPDS_KEYID_LEN + ZPDS_NODE_LEN )) continue;
 			KeyTypeE kp = (KeyTypeE) std::stoull(key.substr(0,ZPDS_KEYID_LEN),0,16);
 			if ( kp >= K_LOGNODE) continue;
+
 			switch(kp) {
 
 			default:	{
-				std::cout << "case K_NOKEY: " << kp << std::endl;
+				// std::cout << "case K_NOKEY: " << kp << std::endl;
 				break;
 			}
-			case K_LOGNODE:	{
-				::zpds::store::TransactionT x;
-				bool status = x.ParseFromString(it->value().ToString());
-				if (status && x.id()>0) std::cout << "case K_LOGNODE:" << std::endl << x.DebugString() << std::endl;
-				break;
-			}
-			case K_PROFILE:	{
-				::zpds::store::ProfileT x;
-				bool status = x.ParseFromString(it->value().ToString());
-				if (status && x.id()>0) std::cout << "case K_PROFILE:" << std::endl << x.DebugString() << std::endl;
-				break;
-			}
-			case K_SIMPLETEMPLATE:	{
-				::zpds::store::SimpleTemplateT x;
-				bool status = x.ParseFromString(it->value().ToString());
-				if (status && x.id()>0) std::cout << "case K_SIMPLETEMPLATE:" << std::endl << x.DebugString() << std::endl;
-				break;
-			}
-			case K_LOOKUPRECORD:	{
-				::zpds::store::LookupRecordT x;
-				bool status = x.ParseFromString(it->value().ToString());
-				if (status && x.id()>0) std::cout << "case K_LOOKUPRECORD:" << std::endl << x.DebugString() << std::endl;
-				break;
-			}
-
+			ZPDS_CASE_MACRO(TransactionT,K_LOGNODE)
+			ZPDS_CASE_MACRO(TagDataT,K_TAGDATA)
+			ZPDS_CASE_MACRO(CategoryT,K_CATEGORY)
+			ZPDS_CASE_MACRO(ExterDataT,K_EXTERDATA)
+			ZPDS_CASE_MACRO(UserDataT,K_USERDATA)
+			ZPDS_CASE_MACRO(LocalDataT,K_LOCALDATA)
+			ZPDS_CASE_MACRO(WikiDataT,K_WIKIDATA)
 			}
 		}
 	}
@@ -94,6 +97,8 @@ public:
 
 } // namespace store
 } // namespace zpds
+
+#undef ZPDS_CASE_MACRO
 
 /** main */
 int main(int argc, char *argv[])
